Fixed: Gaps in cave walls.
Made the first IF statement in the place_wall_in_corner() method more specific. It now requires the below and to the left and the below and to the right neighbours to be rock tiles, on top of just the below neighbour. This fixes a bug where, in some situations, there were gaps showing due to the wrong wall tile being placed.