Updated minimap player pos to work with new tilemap system
As wall tiles no longer take up a full tile, the player's tile position can be seen as inside a wall and the minimap would override wall tiles with the player tile. The minimap now avoids this, however, it could be improved.

This system should be improved when reworking the minimap, which is already planned.